refactor: Consolidated type declarations in types.ts for better organization
