+++ types.ts
@@ -1,0 +1,17 @@
+export type Anime = {
+  date: string;
+  title: string;
+  episode: number;
+  url: string;
+};
+
+export type Book = {
+  title: string;
+  readDate: string;
+  url: string;
+};
+
+export type Movie = {
+  title: string;
+  url: string;
+};
